调整

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/system/SysMemberController.java
@@ -94,7 +94,7 @@ public class SysMemberController extends BaseController {
 
     @PreAuthorize("@ss.hasPermi('system:member:save')")
     @GetMapping("/echo")
-    public SysMember echo(Integer id){
+    public SysMember echo(Long id){
         if(id == null){
             throw new RuntimeException("id不能为空！");
         }
