+++ ruoyi-system/src/main/java/com/ruoyi/system/service/SysMemberService.java
@@ -24,7 +24,7 @@ public interface SysMemberService {
 
     void batchAdd(List<SysMember> members);
 
-    SysMember echo(Integer id);
+    SysMember echo(Long id);
 
     void deduction(SysMember member);
 }
